refactor: 예외 추가 및 예외 슈퍼클래스 디렉터리 이동
