+++ src/main/java/com/psj/itembrowser/security/common/exception/TokenException.java
@@ -1,5 +1,7 @@
 package com.psj.itembrowser.security.common.exception;
 
+import com.psj.itembrowser.security.common.exception.ancestor.CustomRuntimeException;
+
 import lombok.Getter;
 
 /**
