+++ src/main/java/com/psj/itembrowser/security/common/exception/ancestor/CustomAuthenticationException.java
@@ -1,6 +1,8 @@
-package com.psj.itembrowser.security.common.exception;
+package com.psj.itembrowser.security.common.exception.ancestor;
 
 import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
+
+import com.psj.itembrowser.security.common.exception.ErrorCode;
 
 import lombok.Getter;
 import lombok.NonNull;
